Fixing a lot of issues caused by translated planet names.
These are temporary "ugly" fixes, before the KDE 4.2 release that we
might want to revert and re-design later.

The trouble is that in the case of Sun, Earth and Moon, the name()
method returns the English name whereas on other planets, it returns
the translated name (while that's not the way it should be, according
to the documentation in skyobject.h). A quick fix for this is to make
sure that we always use the translated name to find the planet.

There will still be (minor) issues with languages that have a
translation for Pluto.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=901758