Backporting further fix to bug 179034. Quoting the original commit log:
----

Fixing (further) problems with the FOV editor. The localized versions
would accept FOVs that were formatted in the localized format (60,00
in German, for instance) but wouldn't work. This was because the same
bug had to be fixed in two other places.

The FOV editor now seems to work correctly.

----

CCMAIL: kstars-devel@kde.org


svn path=/branches/KDE/4.1/kdeedu/kstars/; revision=906170