Backporting fix from Rev 908470
Original commit log reads as follows:

Fix trouble with links of translated planet names.

This is again, one more of those temporary translation fixes that we
are doing for KDE 4.2. Really, we should revamp the way translations
are handled. That'll fix a whole bunch of bugs.

CCMAIL: kstars-devel@kde.org


svn path=/branches/KDE/4.2/kdeedu/kstars/; revision=908997