+ Adding support for comet magnitudes.
KStars can now pick up H and G parameters (absolute magnitude and
slope parameter) from the comets.dat file if present, and use them to
calculate the magnitude of comets.


+ Fixing formula for comet and asteroid magnitudes

The formula used earlier was applicable only to 2P/Encke and was not
general, and I had mistakenly put that in. Now using the generic
formula.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=910996