This should fix bug 189399.
"Jupiter" might be translated in some languages, so putting an i18n on
the findByName call should prevent it from returning NULL and crashing
KStars. This should not introduce new strings as "Jupiter" is
translated elsewhere.

I can't test this fix because I don't have a language that translates
the string 'Jupiter', but it leaves the German and English versions
(that don't translate Jupiter) untouched. Please test this fix and
close the bug if it works.

CCBUG: 189399
CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=975410