In this changeset clone function is implemented. It allows safe copying of objects. E.g. KSPlanet stored as SkyObject* pointer would be copied as KSPlanet and not sliced.
For all subclasses of SkyObject but DeepSkyObject copy
constructors are removed. In most cases they are pretty useless
and could be replaced with default implementations. 

* Changes in copy constructors
 - Trail in TrailObject wasn't copied. Now it's copied
 - const annotation added in DeepSkyObject copy constructor

* Overloaded assignment operator in EclipticPosition
  (skyobjects/ksplanetbase.h) removed. Default
  implementation would be just fine too.

Please note, despite the fact that a lot of files are changed
only conjustion tool is affected by this change. Objects of
SkyObject class family are not copied much.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=977139