Commiting patch from Debian: http://svn.debian.org/viewsvn/pkg-kde/trunk/packages/kdeedu/debian/patches/00_kfreebsd.diff?revision=14624&view=markup
Many thanks to Sune Vuorela for bringing this to my attention.

Original commit log:

Thu May 14 20:50:08 2009 UTC (8 weeks, 4 days ago) by aurel32:

Test on __linux__ is wrong as the behaviour does not depends on the kernel
but on the userland, that is GNU libc, tested by __GLIBC__. Test on 
__FreeBSD__ and __NetBSD__ are correct though as they represent both the 
kernel and the userland.

The patch does not remove __linux__ in the patch to be as conservative as
possible.


CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=996458