Make the image / info links in the Details Dialog work in translated versions.
The details dialog would earlier look for the link using a translated
string, which wouldn't work. This is fixed in the same manner as in
KSPopupMenu. This closes bug 184266.

Ah, and this is the first KDE commit of the Hackfest we're holding in
IITM, Chennai now. I hope this will be the first of many. Many thanks
to the excited and enthusiastic group of participants who helped me
fix this bug!

BUG: 184266
CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=1030187