Backporting fix for missing stars (r1043599) to 4.3 branch
Original commit log:
--------------------
Ensure that stars with unspecified spectral types / partially
specified spectral types are drawn as well. For the time being, all
stars with undefined spectral classes are drawn with the colors
corresponding to spectral class 'A'.

This fixes the bug that Lukas Middendorf brought up on the mailing
list, regarding stars in M 29.

Many thanks to Lukas, who not only pointed out this hard-to-catch bug,
but also gave me enough information to swat it dead real easy!

CCMAIL: kstars-devel@kde.org


svn path=/branches/KDE/4.3/kdeedu/kstars/; revision=1043604