Show name labels for deep-sky objects.
I've tried to implement all the ideas that were posted as comments in
the wishlist report. Thanks for the report, and the ideas.

It looks like the slider for controlling density of DSO labels isn't
working. Other TODOs are - the colors for the labels are not editable
at the moment, if I amn't wrong. I'm keeping them for the next weekend.

CCMAIL: kstars-devel@kde.org
FEATURE: 207798


svn path=/trunk/KDE/kdeedu/kstars/; revision=1043623