Implement Odd-Even fill for convex polygons as described in Ch.14 of the Red Book. To test it, uncomment the "#define MAKE_KSTARS_SLOW" line.
We disable the colour buffer, then draw triangle fans that invert the stencil 
buffer. At the end any pixel in even-ly-many triangles is eliminated and any in
odd-ly-many triangles is kept. Then we draw a big polygon using this stencil.

While this temporarily solves the non-concavity issue, it does so at the price 
of killing the framerate: when the Milky Way is enabled, the rendering is twice
as slow -- previously, enabling it had no noticable effect on speed.
This is because of all of the extra triangles it has to draw.

Ideally we would change the data to have only concave polygons, but I'm not 
sure of the format, or whether we will ever want to have other convex data.

CCMAIL:kstars-devel@kde.org


svn path=/branches/kstars/hdevalence/kstars/; revision=1153435