+ Draw the horizon line if the ground is selected.
+ Disable the horizon line checkbox if the Opaque Ground checkbox is
  checked.

Earlier, the ground was drawn only if Options::showHorizon() was
true, irrespective of Options::showGround().

While it would be more appropriate to gray out the Show Opaque Ground
checkbox if Horizon Line is turned off, that would be programmatically
more complicated. Hence the not-so-intuitive approach of graying out
Horizon Line if Show Ground is turned on.

I invite suggestions on how this can be made more intuitive.

+ Invoke the slots that disable the appropriate options from the
  constructor, in the Guides section of the Options dialog. Earlier,
  the options were not disabled on frame load.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=1155112