Adding a Generic calendar widget.
Now, this is a bad idea, but there seems to be no alternative at the
moment.

The issue is that we need a KDatePicker-equivalent that handles
customized date tables subclassed from KDateTable (such as the one
we've tweaked to draw the moon phases in the cells), but KDatePicker
does not permit you to access the KDateTable it uses.

While the (probably) ideal way to solve this problem is to get
KDatePicker to support customized KDateTables (I've requested this
with John Layt), this is a make-do thing till that happens. I
implement a "dirty" KDatePicker-replacement around a custom
KDateTable. Support for all calendars other than Gregorian is broken.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=1156270