Filter away points below the horizon in Equatorial mode.
We don't call EquatorialToHorizontal first because it is
expensive and has usually already been called.

Ideally there would be a way to know whether or not
the given SkyPoint had its two sets of coordinates syncronized,
but it doesn't seem that there is a way to do this.

Instead, we add the proviso that the points must have already
had EquatorialToHorizontal called on them.

CCMAIL:kstars-devel@kde.org

svn path=/branches/kstars/hdevalence/kstars/; revision=1160337