Google Code-in task solved by Cezar Mocan, consisting of code refactoring in the ObservingList class.
He removed significant parts of duplicated code and also identified a bug in the trunk (the context menu
was not showing for object on the session list) and fixed it pretty fast, even though he was not required to do so.

Results: cleaner and efficient code, one bug solved.
CCMAIL: kstars-devel@kde.org, akarsh.simha@kdemail.net, mocancezar@gmail.com


svn path=/trunk/KDE/kdeedu/kstars/; revision=1200806