Temporary fix of KStars, just in case I don't fix everything before the 4.6 release.
If this is done, the only significant changes, if I remember, in 4.6
are fixing a lot of projection-related bugs, refraction code, moon
phase almanac and some cities.dat fixes. Maybe there are a couple
more, but I'm disabling OpenGL support.

Hopefully, before the final release, I should have an option to switch
to the "Beta" OpenGL rendering mode which will be a regression from a
feature POV, but will run much much faster, so that many users will
still be interested in that.

CCMAIL: kstars-devel@kde.org, hdevalence@gmail.com


svn path=/branches/KDE/4.6/kdeedu/kstars/; revision=1210036