Not very clean work-around for the InfoBox problem.
Make the Infoboxes subordinate to the widget that does the actual
drawing, and not the QGraphicsView. This is what was, in principle
being done before -- the SkyMap widget had these widgets as
children. But this is not neat, IMO.

Probably, the right way to do this would be to make the InfoBoxes a
child of the QGraphicsView. But then, there are problems with
capturing mouse events.

svn path=/trunk/KDE/kdeedu/kstars/; revision=1210559