The same fseek bug existed in the actual code that reads the data files as well. There was a boundary in Triangulum, on one side of which, faint stars were missing.
This was because KDE_fseek would accept a signed long which can range
to 2^31 - 1 (I hope I've got my basics right), but we want to seek
farther than that. So fseek would fail because it would interpret our
large positive offset as a negative offset.

A solution, as I learned from my past self, is to do the fseek in two
steps. Implementing that solution fixes the Triangulum problem:
http://i.imgur.com/IQsmV.png (Top = Before, Bottom = After)