Fix bug that causes KStars to crash (segmentation fault) in Options window, bef$
Patch that fixes KStars crashing when Options window was closed, before TLEs ar$
Crash was in OpsSatellites::slotItemChanged(), because before TLEs are updated,$
of item which was accessed was 0x0. This patch also fixes view's header (it was$
because model was cleared after header was set) and makes items read-only.

CCBUG: 279028