Fix wrong Daylight Savings Time at KStars startup
This patch fixes an old bug. It turns out that DST (Daylight Savings Time)
setting was not saved and that resulted with no DST correction (fallback
"--" DST rule when all other location information is correct) when
the program was run again. The workaround was to enter Settings->Geographic
and click OK.

The patch adds change of the DST field in Option when the location is changed,
that leads to the new value to be saved.

In the case of a configuration without DST saved, the configuration will be
fixed next time the location is changed or the user enters Settings->
Geographic and clicks OK.

CCBUG:209647