Fixing the conjunction predictions for long time durations. The step size would be way too large if one was looking for conjunctions, of say, Jupiter, over several 1000 years (which is what I was doing when I ran into this).
This is still suboptimal code. I'd appreciate some analysis on this
and some ways to improve this code as well as the initial guess for
the step. It should not be overtly hard, since they are mostly
quasi-periodic stuff.

I want everyone to look at my ugly code (and fix it), so I'm CCing the
ML.
CCMAIL: kstars-devel@kde.org