Deleting broken AAVSO feature.
This feature had been broken for quite some time, so we've decided to
remove it. Due to the big changes at AAVSO site, it might be quite hard
to get it running in the future. Fortunately, we have one person who is
working on it right now.

CCMAIL: kstars-devel@kde.org