Some improvements to the Observation Planner UI.
The date is now picked by a KDateComboBox, which allows for a calendar
popup. The hope was to make that a Moon Phase Calendar popup, but that
requires almost a reimplementation of the KDateComboBox, and it might
not be worth the effort. The internal KDatePicker / KDateTable is not
protected, and hence cannot be replaced using our custom KDateTable
implementation.

Also expanded the abbreviation "WUT". Reorganized some of the buttons.

There's still scope for improvement in the UI, though. The buttons
take too much room. It is possibly better handled using drop-down
menus instead of buttons.