Make sure SkyMap gets to track the mouse, so it can update the coordinates in the status bar.
Firstly, many thanks to ozzzy on IRC (##astronomy on irc.freenode.org)
for pointing out this bug, and pursuing its case with me.

"InfoBoxes" is a transparent "container" widget occupying the entire
screen area, and having the individual InfoBox widgets as
children. InfoBoxes is a child of the painting widget (SkyMapQDraw or
SkyMapGLDraw as the case may be), which is in turn a child of SkyMap.

Although InfoBoxes and SkyMapFooDraw do not need to track the mouse,
they need to have mouseTracking enabled, because no mouse move events
will be initiated and passed up the stack otherwise.

There is a Qt::WA_TransparentForMouseEvents widget attribute, but
this, according to the documentation, blocks mouse events to the child
InfoBox widgets as well.

This is not optimal -- any overlaying widgets that are translucent
would need to have mouse tracking set to true in order to pass the
mouse move events up to the SkyMap.

Is there a better solution?

CCMAIL: kstars-devel@kde.org