Render stars as vectors whenever possible, when exporting sky image.
Stars are not drawn as vectors while rendering the SkyMap, for
efficiency. Bitmap stars render faster on the SkyMap.

Stars are also not drawn as vectors when using the "Real Colors" mode,
because it is harder to render those stars as vectors, although not
impossible. This is a future feature idea.

FEATURE