Initializing the (ecliptic) longitude of perihelion of the earth. Data is from: http://nssdc.gsfc.nasa.gov/planetary/factsheet/earthfact.html Hopefully, we can believe the GSFC :)
The aberration computation relies on this, because the aberration
ellipse (as I understand it) is a downscaled "copy" of the earth's
orbit. Seemingly, this was never initialized, so the orientation of
the ellipse was off by 103 degrees!

Currently, we set it to the constant value from J2000.0, but it does
change with time, mostly because of perturbations from the gravity of
other planets (most notably Jupiter) and slightly because of general
relativity.

According to Wikipedia:
https://en.wikipedia.org/wiki/Apsidal_precession#Long-term_climate
the perihelion completes one revolution in 112,000 years, which leads to a value of 11.6 arcseconds / year.
So the orientation of the aberration ellipse is off by about 11.6 arcseconds a year. This is 5.6e-5 radians.

Accordint to Wikipedia on aberration:
https://en.wikipedia.org/wiki/Aberration_of_light and according to
ksnumbers.cpp:157, the constant of aberration is about 20.5
arcseconds, and Wikipedia explains that this is the "radius" of the
approximate circle traced by the stars.

The annual correction to the coordinates of a star due to the
precession of earth's perihelion would be, in the worst case (assuming
linear term in a Taylor expansion of the trigonometry involved), about
(5.6e-5)*0.0167*20.5 arcseconds / year = 1.83e-5 arcseconds/year.
(The 0.0167 is the eccentricity of the earth. Any corrections should
be in the worst-case linear in the eccentricity.)

If we want to be accurate for 10,000 years, we may still get away
without accounting for this, but if our order of magnitude estimate
above is incorrect over those longer timescales, it may be comparable
to typical atmospheric seeing.

But this should be fine for now! Unless we can find the algorithm for
this... :D

Should be backported to 4.10 and cherry-picked into master.

CCMAIL: kstars-devel@kde.org
(cherry picked from commit 412f1fc462468c5cfa0661fddc1d958b564e5f66)