+ Adding a bunch of fields for Deep-Sky and generic objects.
+ Changing the way magnitudes are interpreted. The old "Magnitude"
  field is now the sortingMagnitude, which is used for determining
  whether the object is visible or not, etc. It need not be scientific
  / accurate / reliable data. We could, in principle, use some
  "handwavy" guesses for this field. (eg: if the blue magnitude is
  available, we may "guess" the visual sorting magnitude by guessing
  an appropriate B-V color index. See, eg DeepStarComponent)

  The real Magnitudes are now stored in a Hash. Those are the values
  that are fit for real display.

The question of StarObject:

StarObject already has support for B and V magnitudes separately. The
question is, should we migrate that to the hash? We have many stars --
not sure if it is optimal to migrate the star magnitudes into the
magnitude hash. Besides, stars are replicated using the placement new
operator, so I'm unsure whether that would lead to any additional
complications.

Any ideas from the list?

We could just override mag( band ) in StarObject if necessary.

CCMAIL: kstars-devel@kde.org