Reorganizing and "optimizing" the KSNumbers::updateValues code.
Initially, I made the mistake of thinking that this rather suboptimal
code was in the innermost loop. By the time it dawned on me that it
was kind-of immaterial, I'd already improved it. Anyway, this is more
clean in many other ways too.

Hopefully, it doesn't introduce any bugs. I did check the code twice
just to make sure.