Fix crash while exporting image
The crash in setLegendAlpha() while exporting images, introduced when
I separated the ExportImageDialog dialog from the ImageExporter
"backend", was (most likely) caused because of an uninitialized
ImageExporter pointer. This caused ImageExporter to never be
constructed, but used, naturally leading to a crash. This should
hopefully fix the bug. More testing is necessary, but preliminary
testing shows that it works.

CCMAIL: kstars-devel@kde.org