Make supernovae update process asynchronous.
Make the supernova update process asynchronous so that it doesn't slow
down KStars startup even if the network connection is down.

To do this, the update method was split into two methods, one to
trigger the update, and the other to reload the data after the update
process is finished.