Show asteroids and comets by default.
It looks like magnitude estimation is working fairly well for comets
and asteroids, so there seems to be no harm in enabling comet and
asteroid display by default.

I'm not sure of this, though, and so I'm copying the mailing list to
see if there's any dissent / feedback on this.

CCMAIL: kstars-devel@kde.org