Revert the default projection to "Lambert Equal-area Azimuthal"
Earlier, I changed the default projection to Gnomonic because it looked
like one was "inside" the star globe. On second thoughts, the distortion
is much higher in Gnomonic, so I'm reverting back to Lambert.

I'm not sure what is the best preference for most users. I invite
suggestions and feedback on this subject.

CCMAIL: kstars-devel@kde.org