Move image and info links in the popup menu into dedicated submenus.
The popup menu for some objects would end up becoming too long because
of the number of information and image links. So moving them into
dedicated submenus helps keep the context menu short and improves
usability. Besides, grouping the links as image / info resources helps
users find them more easily.

I decided not to put the DSS / SDSS imagery under the image links
submenu, because they are available for any point on the sky (within
the footprints of the surveys). If we ever create a submenu for sky
surveys, it should possibly go in there.

CCMAIL: kstars-devel@kde.org