Prevent potential segfault in observing list.
While this does not really find and eliminate the source of the null
object, it at least prevents KStars from crashing while trying to add
objects to the observing list. The segfault can result in severe loss
of observing wishlist data because the data file is overwritten each
time and the data save could be aborted by a segfault.

It happened to me today, and I have to redo 2 hours of scouring
through observing lists.

Will not backport as we're going to release 4.11 soon.

BUG