Allow standard keypress events in the tables in the Observation Planner
The table showing objects in the observation planner had the
keypresses eaten by the observation planner class. While only the
"delete" key event was actually handled, this meant that the standard
keybindings for navigation did not work. Most notably, the arrow keys
for navigation would not work, which made browsing through objects
rather difficult.

This fixes it by absorbing only the delete key event.

Furthermore, it seems like the event filter may not be necessary. Any
possible reasons for preferring it over regular signal-slot
connections must be investigated and the former replaced with the
latter if no reasons exist.

BUG