Fix twilight display on the altitude vs time plot in the observing list.
The dawn and dusk (astronomical twilight) times were not set in the
AVT Plot Widget that is embedded within the observing list UI. There
is a lot of code duplication already, and while this adds more, it
fixes the problem from the user's point-of-view.

Hopefully, this code duplication can be removed after Henry de
Valence's ksengine becomes a reality.

CCMAIL: hdevalence@hdevalence.ca