Add recognition of triangular patterns to the star-hopper
Some improvements to the starhopper and some bugfixes (eg: the change
involving looking for asterisms around the star-hop destination rather
than source!)

But primarily, adding a rather conservative triangle-identification
feature, that identifies various kinds of triangles. The kind of
triangle identified is printed in the debug output of the
star-hopper. All this is Beta, obviously.

CCMAIL: kstars-devel@kde.org