Initialize Rearth in KSPlanetBase to NaN
The member Rearth describes the distance to earth from the
planet. This was initialized to zero. It is much cleaner to initialize
it to NaN.

This may introduce a number of crashes, especially in debug mode, but
it will help expose computation bugs that would otherwise be hard to
find.

CCMAIL: kstars-devel@kde.org