Fix infinity in accounting for General relativity effects.
Rearth of KSPlanetBase initialized to 0 caused crashes sometimes when
GR effects were involved. The place where I encountered this problem
was in the observing wishlist. When the observing wishlist is
initialized, KSSun's geocentric position has not been recomputed, and
this led to a number of crashes. So we work around the situation by
doing an approximation -- we just declare the earth-sun distance to be
1 AU. Clearly, this is not robust and could lead to miscalculations,
but GR effects is such a niche, that the only purpose is to look at
the sun and toggle them on and off to simulate Eddington's
observations. Therefore, I would deem it inconsequential that this
code can give you the wrong result at times. In any case, the
discrepancy, if any is observed, should be small.

This fixes crashes in debug mode while loading the observing wishlist.