Adding an "eyepiece view" feature to the observation planner
This is a very beta feature at the moment. It renders a view of the
eyepiece and DSS / other imagery if available, side-by-side. The view
can be rotated, flipped and inverted to help match the view through a
telescope / camera.

The DSS image and the sky map view are oriented correctly only when
using equatorial coordinates. One needs to work out some (presumably
well-known, standard) trigonometry to rotate the DSS in horizontal
coordinates.

The feature is very buggy, but works. Will polish it if and when I
have the time. Please feel free to contribute.

FEATURE
CCMAIL: kstars-devel@kde.org