Update the missing file message box to reflect changes in standard dirs.
The old standard dirs are invalid with KF5, and so the static list of
standard dirs (which was, for starters, not cross-platform) has been
replaced by the actual list that changes from
platform-to-platform. However, this is still not accurate since not
all files that this message box shows up for need to be located in a
QStandardPaths::DataLocation, although most will be.

CCMAIL: mutlaqja@ikarustech.com