Add option to toggle display of deep-sky objects of unknown magnitude
So far, objects that had unknown magnitudes were never drawn in
KStars, except for a workaround to permit the display of IC objects,
many of which had unknown magnitudes.

While the Steinicke catalog is revised, we use the blue magnitudes
from it, and not all objects have blue magnitudes. So using the
Steinicke catalog renders many objects not drawn.

This commit adds an option in the Catalog settings UI that allows the
user to toggle the display of objects of unknown magnitudes. The
default state of this setting is 'true', so that the user is not
frustrated to find some NGC / IC objects invisible. The patch also
removes the work-around for IC objects.

This only affects objects of unknown magnitude -- the faint objects
are still hidden and revealed as the user zooms in.

BUG:
GUI: