Rename variables in the Observing List code to more intuitive names
The observing list in KStars started out as just one list. We now have
the ability to manage a wishlist and a session plan. When the session
plan was added, the old variables were not clearly renamed to indicate
the difference. This commit renames the variables for
clarity. Building has been checked and KStars builds and executes
okay.

The following are the changes:

m_Model -> m_WishListModel
m_Session -> m_SessionModel
m_SortModel -> m_WishListSortModel
m_SortModelSession -> m_SessionSortModel