Add a few private methods to ObservingList to reduce code duplication
At any give time, the ObservingList class manages two MVC lists -- the
wishlist and the session plan. A lot of methods operate on the
'active' model / list, which corresponds to the list currently being
seen in the UI.

This has led to a lot of code duplication.

In this commit, I add a few methods to return the active list, model,
view etc. so that code duplication is reduced. This is not the way to
go, but it seems to me to be a good first step to reduce confusion to
do further refactoring.