Refactor observing list popup-menu and simplify code in calling it.
Note: I'm copying the list because testing may be necessary. The patch
is hopefully not controversial.

The observing list popup menu took many parameters to decide which
options were displayed and which were not. It is much more simple (and
this can be changed if future needs differ) to accept fewer flags that
tell the menu the context in which it is being called -- i.e. were
multiple objects selected? were we in session view or wishlist view?
is INDI available?

This commit implements these changes and also simplifies the code that
calls the popup menu.

Importantly, this prevents a crash which involved calling "Eyepiece
View (Beta)" on multiple objects. Eyepiece view only operates on a
single object, and hence, it is now not possible to call Eyepiece view
on multiple objects.

CCMAIL: kstars-devel@kde.org