In the Geographic Location form, inform users of validation errors.
Daniel just reported a bug on the bugzilla about a usability issue in
the Geographic Location form.

The issue is that users are not informed of validation errors, so it's
not obvious as to why the 'Add City to List' button is grayed
out. (Also, counterintuitively, the longitude box comes before the
latitude box. Maybe this should be changed too...).

This commit fixes the problem by including a label that displays the
appropriate error message.

Thanks, Daniel! I was looking for some low-hanging fruit :)

BUG: 341965
GUI: