Prelimiary versions of UI to display the results from the Star Hopper.
The results of the star hopper are displayed using a UI that just
lists the star at each step of the star-hop and allows easy access to
details of the star and also makes it easy to center the map on each
step of the star-hop.

(Commit log authored by Akarsh Simha. Code / patch by Utkarsh Simha)

REVIEW: 121176

CCMAIL: utkarshsimha@gmail.com