Adding the HTMesh perl wrapper used by data-manipulation scripts
Data manipulation scripts used to create binary star data files are
available in the kstars/kstars/data/tools directory. Some of the perl
scripts rely on a (third party) perl wrapper to the HTMesh library,
which was missing. It has now been included in the KStars sources for
use with these scripts.

I also added some extra information to the README and updated my
e-mail address.