Adding star-hop instructions to the star hopper UI + other small changes
Directions previously printed to the debug console are now displayed
on the UI.

Also, some variable naming conventions have been changed, and
indentation faults have been fixed.

The buttons on the UI are now placed in a horizontal layout and a
label to display the aforementioned directions is added.

CCMAIL: utkarshsimha@gmail.com