Fix image display in the observation planner
Must use QUrl::fromLocalFile to create proper URLs from file names.

Not having done this resulted in the image previews not being
displayed. This commit fixes the same.

(This issue was introduced during KF5 transition.)

Also, we depend on the ThumbCreator service to show the previews. It
might be a good idea to remove this dependency using QLabel instead,
like we do in the detail dialog. On my laptop, the delay in scaling a
large DSS image (M 42) in QLabel (as seen in the detail dialog) seems
to be negligibly small, so having cached thumbnails might be
superfluous. This will also ensure consistency across various parts of
KStars.

The alternative is that must at the very least provide the user with
some warning that (s)he might need to install additional packages.