Show date + time on solar system object trails.
This needs to be made configurable and better in general, but we now
draw labels indicating the date and time along the trail of a solar
system object (anything that is derived from KSPlanetBase).

This means, we can plot comet trajectories with date + time
annotations!  Example for Comet 2014 Q2 Lovejoy, which is currently
bright and sporting a nice tail:
http://wstaw.org/m/2015/01/18/2014Q2.gif

FEATURE
CCMAIL: kstars-devel@kde.org
GUI
DIGEST