Be more conservative in resetting RA0 and Dec0
With dms::reduce() and calculations and the like, it is not entirely
unlikely that we might have a negative RA. Be more conservative and
reset RA0, Dec0 only if they are NaN, or if |Dec| is larger than 90.0
degrees.

This might have led to some issues that might be fixed now.